refactor: refactor main function
