+++ internal/cmd/history.go
@@ -1,0 +1,38 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/yuluo-yx/typo/internal/config"
+	"github.com/yuluo-yx/typo/internal/engine"
+)
+
+func cmdHistory(args []string) int {
+	if len(args) < 1 {
+		fmt.Fprintln(os.Stderr, "Error: subcommand required (list, clear)")
+		return 1
+	}
+
+	cfg := config.Load()
+	h := engine.NewHistory(cfg.ConfigDir)
+
+	switch args[0] {
+	case "list":
+		entries := h.List()
+		for _, e := range entries {
+			fmt.Printf("%s -> %s (used %d times)\n", e.From, e.To, e.Count)
+		}
+		return 0
+	case "clear":
+		if err := h.Clear(); err != nil {
+			fmt.Fprintf(os.Stderr, "Error: %v\n", err)
+			return 1
+		}
+		fmt.Println("History cleared")
+		return 0
+	default:
+		fmt.Fprintf(os.Stderr, "Unknown subcommand: %s\n", args[0])
+		return 1
+	}
+}
